Drop distinction between mysql and mysql2
The `mysql` gem has been deprecated in favor of the `mysql2` gem.  At this point, we shouldn't make a distinction and just point people to the `mysql2` gem.

This also removes a warning about MySQL in Snow Leopard, which included a typo that has been there for about a year.  Given the length of time the typo existed, I would say we're in the clear to remove the entire warning.